Update to use Deno more
