+++ src/legacy/editor-service/common/EditorProcess.ts
@@ -1,7 +1,7 @@
 import { ChildProcessByStdio } from "child_process";
 import { Readable, Writable } from "stream";
 import { TextDecoder } from "util";
-import { DprintExecutable } from "../../../DprintExecutable";
+import { DprintExecutable } from "../../../executable/DprintExecutable";
 import { Logger } from "../../../logger";
 
 const textDecoder = new TextDecoder();
